requirements.txt
argparse
datetime
matplotlib
numpy
pillow
quadprog
torchvision
urllib3
uuid
